EKIR-775 Move FeedbackForm to forms module
