+++ app.py
@@ -9,19 +9,7 @@ from datetime import datetime
 from flask import request, render_template, redirect, url_for
 
 from flask_bootstrap import Bootstrap5
-
-from flask_wtf import FlaskForm, CSRFProtect
-from wtforms import (
-    StringField,
-    HiddenField,
-    TextAreaField,
-    EmailField,
-    SelectField,
-    SubmitField,
-    validators,
-    ValidationError
-)
-
+from flask_wtf import CSRFProtect
 from flask_babel import lazy_gettext as _
 from flask_babel import Babel
 
@@ -47,52 +35,6 @@ bootstrap = Bootstrap5(app)
 # Flask-WTF requires this line
 csrf = CSRFProtect(app)
 app.secret_key = secrets.token_urlsafe(16)
-
-def validate_municipality(form, field):
-    if field.data == "":
-        raise ValidationError(_("Please select a valid municipality."))
-
-class FeedbackForm(FlaskForm):
-    subject = SelectField(
-        _("Subject"),
-        choices=[
-            (_("General feedback")),
-            (_("Material procurement")),
-            (_("Material error")),
-            (_("Technical problem")),
-            (_("Other")),
-        ],
-    )
-    device_manufacturer = HiddenField(_("Manufacturer"), [validators.Optional()])
-    device_model = HiddenField(_("Device model"), [validators.Optional()])
-    version_name = HiddenField(_("Software version name"), [validators.Optional()])
-    version_code = HiddenField(_("Software version code"), [validators.Optional()])
-    commit = HiddenField(_("Commit"), [validators.Optional()])
-    book_name = StringField(
-        _("Book name"), [validators.Optional(), validators.Length(1, 128)]
-    )
-    message = TextAreaField(
-        _("Message"),
-        [validators.DataRequired(), validators.Length(1, 2048)],
-    )
-    
-    municipality = SelectField(
-        _("My home municipality that receives this feedback"),
-        choices=[("", _("Select a municipality"))] + indexed_municipalities(),
-        render_kw={
-            "class": "form-select",
-            "data-control": "select2",
-            "data-dropdown-parent": "body",
-        },
-        validators=[validate_municipality]
-    )
-    email = EmailField(
-        _("Email address, if you want an answer to your feedback (Optional)"),
-        [validators.Optional()],
-    )
-    submit = SubmitField(
-        _("Send"),
-    )
 
 @app.route(root_path, methods=["GET", "POST"])
 def feedback(name=None):
